adjust balance and setting

+++ Classes/Gacha.cpp
@@ -118,7 +118,7 @@ void Gacha::setNewGacha(WorldInfo* info)
     document.Parse<0>(jsonStr.c_str());
     rapidjson::Value& gachaDoc = document[std::to_string(info->gachaId).c_str()];
     _least = gachaDoc["least"].GetInt();
-    _price = (long int)gachaDoc["price"].GetDouble();
+    _price = (long long int)gachaDoc["price"].GetDouble();
     rapidjson::Value& listDoc = gachaDoc["list"];
     for (int i = 0; i < listDoc.Size(); i++) {
         rapidjson::Value& v = listDoc[i];
@@ -139,7 +139,7 @@ void Gacha::setNewGacha(WorldInfo* info)
     }
 }
 
-long int Gacha::getPrice()
+long long int Gacha::getPrice()
 {
     return _price;
 }
